+++ src/knx/association_table_object.h
@@ -6,7 +6,7 @@ class AssociationTableObject : public TableObject
 {
   public:
     AssociationTableObject(Memory& memory);
-    void readProperty(PropertyID id, uint32_t start, uint32_t& count, uint8_t* data);
+    void readProperty(PropertyID id, uint16_t start, uint8_t& count, uint8_t* data);
     ObjectType objectType() { return OT_ASSOC_TABLE; }
 
     uint8_t* restore(uint8_t* buffer);
