+++ src/knx/cemi_server_object.h
@@ -5,8 +5,8 @@
 class CemiServerObject: public InterfaceObject
 {
 public:
-    void readProperty(PropertyID id, uint32_t start, uint32_t& count, uint8_t* data) override;
-    void writeProperty(PropertyID id, uint32_t start, uint8_t* data, uint32_t& count) override;
+    void readProperty(PropertyID id, uint16_t start, uint8_t& count, uint8_t* data) override;
+    void writeProperty(PropertyID id, uint16_t start, uint8_t* data, uint8_t& count) override;
     uint8_t propertySize(PropertyID id) override;
     uint8_t* save(uint8_t* buffer);
     uint8_t* restore(uint8_t* buffer);
