+++ src/knx/device_object.cpp
@@ -2,7 +2,7 @@
 #include "device_object.h"
 #include "bits.h"
 
-void DeviceObject::readProperty(PropertyID propertyId, uint32_t start, uint32_t& count, uint8_t* data)
+void DeviceObject::readProperty(PropertyID propertyId, uint16_t start, uint8_t& count, uint8_t* data)
 {
     switch (propertyId)
     {
@@ -59,7 +59,7 @@ void DeviceObject::readProperty(PropertyID propertyId, uint32_t start, uint32_t&
     }
 }
 
-void DeviceObject::writeProperty(PropertyID id, uint32_t start, uint8_t* data, uint32_t& count)
+void DeviceObject::writeProperty(PropertyID id, uint16_t start, uint8_t* data, uint8_t& count)
 {
     switch (id)
     {
