+++ src/knx/device_object.h
@@ -7,8 +7,8 @@
 class DeviceObject: public InterfaceObject
 {
 public:
-    void readProperty(PropertyID id, uint32_t start, uint32_t& count, uint8_t* data);
-    void writeProperty(PropertyID id, uint32_t start, uint8_t* data, uint32_t& count);
+    void readProperty(PropertyID id, uint16_t start, uint8_t& count, uint8_t* data);
+    void writeProperty(PropertyID id, uint16_t start, uint8_t* data, uint8_t& count);
     uint8_t propertySize(PropertyID id);
     uint8_t* save(uint8_t* buffer);
     uint8_t* restore(uint8_t* buffer);
