+++ src/knx/group_object_table_object.cpp
@@ -14,7 +14,7 @@ GroupObjectTableObject::~GroupObjectTableObject()
     freeGroupObjects();
 }
 
-void GroupObjectTableObject::readProperty(PropertyID id, uint32_t start, uint32_t& count, uint8_t* data)
+void GroupObjectTableObject::readProperty(PropertyID id, uint16_t start, uint8_t& count, uint8_t* data)
 {
     switch (id)
     {
