+++ src/knx/group_object_table_object.h
@@ -10,7 +10,7 @@ class GroupObjectTableObject : public TableObject
   public:
     GroupObjectTableObject(Memory& memory);
     virtual ~GroupObjectTableObject();
-    void readProperty(PropertyID id, uint32_t start, uint32_t& count, uint8_t* data);
+    void readProperty(PropertyID id, uint16_t start, uint8_t& count, uint8_t* data);
     ObjectType objectType() { return OT_GRP_OBJ_TABLE; }
     uint16_t entryCount();
     GroupObject& get(uint16_t asap);
