+++ src/knx/interface_object.cpp
@@ -46,14 +46,14 @@ void InterfaceObject::readPropertyDescription(uint8_t& propertyId, uint8_t& prop
     }
 }
 
-void InterfaceObject::readProperty(PropertyID id, uint32_t start, uint32_t &count, uint8_t *data)
+void InterfaceObject::readProperty(PropertyID id, uint16_t start, uint8_t& count, uint8_t* data)
 {
     // Set number of elements to zero as we are in the end of the call chain
     // Nobody processed the property before.
     count = 0;
 }
 
-void InterfaceObject::writeProperty(PropertyID id, uint32_t start, uint8_t *data, uint32_t &count)
+void InterfaceObject::writeProperty(PropertyID id, uint16_t start, uint8_t* data, uint8_t& count)
 {
     // Set number of elements to zero as we are in the end of the call chain
     // Nobody processed the property before.
